requirements.txt
# PyTorch with CUDA 11.8
torch>=2.1.2
torchvision>=0.16.2
--extra-index-url https://download.pytorch.org/whl/cu118

# Other Python packages
edt
h5py
itk
lightning
monai
mrcfile
numpy
pandas
scipy
plotly
PyYAML
scikit_learn
scipy
scikit-image
starfile
umap-learn